Addressing Comments
> In this sentence: “Students should be aware that everybody, regardless of their disability, is able to use web and technology equally.” Isn't this our desired state of the world, not the current? We want everybody to use web and technology equally, but for that, we need people to know how to create accessible interfaces.

Completely agree. I have reformulated this a little bit but I still think we need to keep focus on how to raise awareness that people with disabilities can use the web when interfaces are accessible. Of course weneed people to know how to develop those interfaces but this knowledge will be dealth with in future modules / units.