English correction (minor) plus minor improvements
Changed "headers scope" to "scope of the headers" -- or it could be "headers' scope" or "header scope". I try to avoid using the apostrophe with a plural if possible.
Remove "to tell" in two places as it is not necessary there.
Change "Have students" to "Ask students" which is clearer and consistent with other suggestions to authors.
Clarified what "some screen readers announce" as I felt this was an important statement and worthy of the extra words to make it clear what screen readers do in both situations.
Simplified use of "one another" in reference to distinguishing "a table from other tables".